Page / Context / Browser
Pages open in same browser window but under same context and same browser.
When Context is different but under the same browser, two windows open and we still see one chromium in task bar of the computer.
When Browser is different (hence the context also is different), then we see additional window chromium opened in task bar of the computer.